requirements.txt
python-dotenv
playwright
firebase-admin
google-generativeai
streamlit
